+++ engine/runtime/src/test/scala/org/enso/compiler/test/core/GraphTest.scala
@@ -1,10 +1,6 @@
 package org.enso.compiler.test.core
 
-import org.enso.compiler.core.Core
 import org.enso.compiler.test.CompilerTest
-import org.enso.graph.{Graph => PrimGraph}
 
 class GraphTest extends CompilerTest {
-  import Core._
-  implicit val graph: PrimGraph.GraphData[Core.CoreGraph] = PrimGraph[Core.CoreGraph]()
 }
